+++ apps/web/src/components/mainpage/greetingsection/Missions.tsx
